+++ src/Compute/Compute.Test/ScenarioTests/DiskRPTests.ps1
@@ -1171,6 +1171,99 @@ function Test-SnapshotConfigDiskAccessNetworkPolicy
 
 <#
 .SYNOPSIS
+Testing the Get-DiskEncryptionSetAssociatedResources cmdlet. 
+Creates a DiskEncyptionSet object, 
+then associates a disk to it. 
+Check with the cmdlet,
+add another disk associated to the encryptionset,
+then check with the cmdlet again.
+#>
+function Test-GetDiskEncryptionSetAssociatedResource
+{
+    # Setup
+    $loc = 'eastus';
+    $rgname = 'EncSetAssociTest';
+    $iteration = '4'
+    $vaultName = 'kv'+ $iteration + $rgname ;
+
+    
+    try
+    {
+        # Create keyvault 
+
+        <#
+        # Note: In order to record this test, you need to run the following commands to create KeyValut key and KeyVault secret in a separate Powershell window.
+
+        $kekName = 'key' + $iteration + $rgname;
+        $secretname = 'mysecret00';
+        $secretdata = 'mysecretvalue00';
+        $securestring = ConvertTo-SecureString $secretdata -Force -AsPlainText;
+
+        New-AzResourceGroup -Name $rgname -Location $loc -Force;
+        $vault = New-AzKeyVault -VaultName $vaultName -ResourceGroupName $rgname -Location $loc -Sku Standard -EnablePurgeProtection;
+        $mocksourcevault = $vault.ResourceId;
+        Set-AzKeyVaultAccessPolicy -VaultName $vaultName -ResourceGroupName $rgname -EnabledForDiskEncryption;
+        $kek = Add-AzKeyVaultKey -VaultName $vaultName -Name $kekName -Destination "Software";
+        $secret = Set-AzKeyVaultSecret -VaultName $vaultName -Name $secretname -SecretValue $securestring;
+        $mockkey = $kek.Id
+
+        $sub = Get-AzContext;
+        $subId = $sub.Subscription.id;
+        $mocksourcevault = '/subscriptions/' + $subId + '/resourceGroups/' + $rgname + '/providers/Microsoft.KeyVault/vaults/' + $vaultName;
+
+        # Create DiskEncyptionSet
+        $diskEncryptionSetConfig = New-AzDiskEncryptionSetConfig -Location $loc -KeyUrl $mockkey -SourceVaultId $mocksourcevault -IdentityType "SystemAssigned" 
+        $diskEncryptionSetName = "encryptionSet0" + $iteration;
+        $diskEncryptionSet = New-AzDiskEncryptionSet -ResourceGroupName $rgname -Name $diskEncryptionSetName -inputobject $diskEncryptionSetConfig
+
+        # set access policy for diskencryptionset in keyvault
+        $access = @("get", "wrapkey", "unwrapkey")
+        Set-AzKeyVaultAccessPolicy -vaultname $vaultName -objectId $diskEncryptionSet.Identity.PrincipalId -PermissionsToKeys $access
+
+        #>
+
+        $diskEncryptionSetName = "encryptionSet0" + $iteration;
+        $diskEncryptionSet = Get-AzDiskEncryptionSet -ResourceGroupName $rgname -Name $diskEncryptionSetName;
+        
+        # Create a disk and associate it with EncryptionSet
+        $diskconfig00 = New-AzDiskConfig -Location $loc -Zone "1" -DiskSizeGB 1 -AccountType "Standard_LRS" -OsType "Windows" -CreateOption "Empty" -HyperVGeneration "V1" -diskencryptionsetid $diskEncryptionSet.id -EncryptionType 'EncryptionAtRestWithCustomerKey';
+        $disk00Name = "disk00" + $iteration
+        $disk00 = New-AzDisk -ResourceGroupName $rgname -DiskName $disk00Name -Disk $diskconfig00
+
+        # check association
+        $res = get-AzDiskEncryptionSetAssociatedResource -resourceGroupName $rgname -diskEncryptionSetName $diskEncryptionSetName;
+        Assert-NotNull $res
+        Assert-True { $res.count -eq 1}
+        Assert-AreEqual $res[0] $disk00.id
+
+        # create another disk without disk encryption set association
+        $diskconfig01 = New-AzDiskConfig -Location $loc -Zone "1" -DiskSizeGB 1 -AccountType "Standard_LRS" -OsType "Windows" -CreateOption "Empty" -HyperVGeneration "V1";
+        $disk01Name = "disk01" + $iteration
+        $disk01 = New-AzDisk -ResourceGroupName $rgname -DiskName $disk01Name -Disk $diskconfig01
+
+        # check association (count still 1)
+        $res = get-AzDiskEncryptionSetAssociatedResource -resourceGroupName $rgname -diskEncryptionSetName $diskEncryptionSetName;
+        Assert-True {$res.count -eq 1}
+
+        # Update disk to associate with encryptionset
+        $diskUpdatedConfig = New-AzDiskUpdateConfig -EncryptionType 'EncryptionAtRestWithCustomerKey' -DiskEncryptionSetId $diskEncryptionSet.Id
+        Update-AzDisk -ResourceGroupName $rgname -DiskName $disk01Name -DiskUpdate $diskUpdatedConfig
+
+        # check association (count = 2)
+        $res = get-AzDiskEncryptionSetAssociatedResource -resourceGroupName $rgname -diskEncryptionSetName $diskEncryptionSetName;
+        Assert-True {$res.count -eq 2}
+    }
+    finally
+    {
+        # Cleanup
+        # Clean-ResourceGroup $rgname
+        Remove-AzDisk -ResourceGroupName $rgName -DiskName $disk00Name -Force
+        Remove-AzDisk -ResourceGroupName $rgName -DiskName $disk01Name -Force
+    }
+}
+
+<#
+.SYNOPSIS
 Testing the new parameters 
 Tier
 LogicalSectorSize 
@@ -1240,5 +1333,4 @@ function Test-DiskConfigTierSectorSizeReadOnly
             # Cleanup
             Clean-ResourceGroup $rgname
 		}
-
 }